requirements.txt
ipykernel
langchain
python-dotenv
langchain_groq
fastapi
langserve
uvicorn
sse_starlette
